refactor axios api
