+++ src/views/MyPage.vue
@@ -109,7 +109,7 @@
 
 <script>
 import MainHeaderVue from "@/components/MainHeader.vue";
-import api from "@/api/index";
+import { auth as api } from "@/api";
 import { mapState, mapMutations } from "vuex";
 
 export default {
